requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
babel==2.16.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
holidays==0.66
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
narwhals==1.27.1
nltk==3.9.1
numpy==2.2.1
opcua==0.98.13
openai==1.63.2
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pd==0.0.4
pillow==11.0.0
plotly==6.0.0
prophet==1.1.6
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycoingecko==3.2.0
pydantic==2.10.6
pydantic_core==2.27.2
pyModbusTCP==0.3.0
pyparsing==3.2.1
pyperclip==1.9.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzbar==0.1.9
regex==2024.11.6
requests==2.26.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
stanio==0.5.1
textblob==0.19.0
threadpoolctl==3.5.0
tkcalendar==1.6.1
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
